Split off Solution::Route as pyvrp::Route
