+++ pyvrp/cpp/Solution.h
@@ -4,6 +4,7 @@
 #include "Measure.h"
 #include "ProblemData.h"
 #include "RandomNumberGenerator.h"
+#include "Route.h"
 
 #include <functional>
 #include <iosfwd>
@@ -41,237 +42,6 @@ class Solution
     using Depot = size_t;
     using VehicleType = size_t;
 
-public:
-    /**
-     * Route(data: ProblemData, visits: list[int], vehicle_type: int)
-     *
-     * A simple class that stores the route plan and some statistics.
-     */
-    class Route
-    {
-        using Visits = std::vector<Client>;
-
-        Visits visits_ = {};           // Client visits on this route
-        Distance distance_ = 0;        // Total travel distance on this route
-        Cost distanceCost_ = 0;        // Total cost of travel distance
-        Distance excessDistance_ = 0;  // Excess travel distance
-        Load delivery_ = 0;       // Total delivery amount served on this route
-        Load pickup_ = 0;         // Total pickup amount gathered on this route
-        Load excessLoad_ = 0;     // Excess pickup or delivery demand
-        Duration duration_ = 0;   // Total duration of this route
-        Cost durationCost_ = 0;   // Total cost of route duration
-        Duration timeWarp_ = 0;   // Total time warp on this route
-        Duration travel_ = 0;     // Total *travel* duration on this route
-        Duration service_ = 0;    // Total *service* duration on this route
-        Duration wait_ = 0;       // Total *waiting* duration on this route
-        Duration release_ = 0;    // Release time of this route
-        Duration startTime_ = 0;  // (earliest) start time of this route
-        Duration slack_ = 0;      // Total time slack on this route
-        Cost prizes_ = 0;         // Total value of prizes on this route
-
-        std::pair<double, double> centroid_;  // Route center
-        VehicleType vehicleType_;             // Type of vehicle
-        Depot startDepot_;                    // Assigned start depot
-        Depot endDepot_;                      // Assigned end depot
-
-    public:
-        [[nodiscard]] bool empty() const;
-
-        /**
-         * Returns the number of clients visited by this route.
-         */
-        [[nodiscard]] size_t size() const;
-
-        [[nodiscard]] Client operator[](size_t idx) const;
-
-        Visits::const_iterator begin() const;
-        Visits::const_iterator end() const;
-
-        /**
-         * Route visits, as a list of clients.
-         */
-        [[nodiscard]] Visits const &visits() const;
-
-        /**
-         * Total distance travelled on this route.
-         */
-        [[nodiscard]] Distance distance() const;
-
-        /**
-         * Total cost of the distance travelled on this route.
-         */
-        [[nodiscard]] Cost distanceCost() const;
-
-        /**
-         * Distance in excess of the vehicle's maximum distance constraint.
-         */
-        [[nodiscard]] Distance excessDistance() const;
-
-        /**
-         * Total client delivery load on this route.
-         */
-        [[nodiscard]] Load delivery() const;
-
-        /**
-         * Total client pickup load on this route.
-         */
-        [[nodiscard]] Load pickup() const;
-
-        /**
-         * Pickup or delivery load in excess of the vehicle's capacity.
-         */
-        [[nodiscard]] Load excessLoad() const;
-
-        /**
-         * Total route duration, including travel, service and waiting time.
-         */
-        [[nodiscard]] Duration duration() const;
-
-        /**
-         * Total cost of the duration of this route.
-         */
-        [[nodiscard]] Cost durationCost() const;
-
-        /**
-         * Total duration of service on this route.
-         */
-        [[nodiscard]] Duration serviceDuration() const;
-
-        /**
-         * Amount of time warp incurred on this route.
-         */
-        [[nodiscard]] Duration timeWarp() const;
-
-        /**
-         * Total duration of travel on this route.
-         */
-        [[nodiscard]] Duration travelDuration() const;
-
-        /**
-         * Total waiting duration on this route.
-         */
-        [[nodiscard]] Duration waitDuration() const;
-
-        /**
-         * Start time of this route. This is the earliest possible time at which
-         * the route can leave the depot and have a minimal duration and time
-         * warp. If there is positive :meth:`~slack`, the start time can be
-         * delayed by at most :meth:`~slack` time units without increasing the
-         * total (minimal) route duration, or time warp.
-         *
-         * .. note::
-         *
-         *    It may be possible to leave before the start time (if the depot
-         *    time window allows for it). That will introduce additional waiting
-         *    time, such that the route duration will then no longer be minimal.
-         *    Delaying departure by more than :meth:`~slack` time units always
-         *    increases time warp, which could turn the route infeasible.
-         */
-        [[nodiscard]] Duration startTime() const;
-
-        /**
-         * End time of the route. This is equivalent to
-         * ``start_time + duration - time_warp``.
-         */
-        [[nodiscard]] Duration endTime() const;
-
-        /**
-         * Time by which departure from the depot can be delayed without
-         * resulting in (additional) time warp or increased route duration.
-         */
-        [[nodiscard]] Duration slack() const;
-
-        /**
-         * Earliest time at which this route can leave the depot. Follows from
-         * the release times of clients visited on this route.
-         *
-         * .. note::
-         *
-         *    The route's release time should not be later than its start time,
-         *    unless the route has time warp.
-         */
-        [[nodiscard]] Duration releaseTime() const;
-
-        /**
-         * Total prize value collected on this route.
-         */
-        [[nodiscard]] Cost prizes() const;
-
-        /**
-         * Center point of the client locations on this route.
-         */
-        [[nodiscard]] std::pair<double, double> const &centroid() const;
-
-        /**
-         * Index of the type of vehicle used on this route.
-         */
-        [[nodiscard]] VehicleType vehicleType() const;
-
-        /**
-         * Location index of the route's starting depot.
-         */
-        [[nodiscard]] Depot startDepot() const;
-
-        /**
-         * Location index of the route's ending depot.
-         */
-        [[nodiscard]] Depot endDepot() const;
-
-        /**
-         * Returns whether this route is feasible.
-         */
-        [[nodiscard]] bool isFeasible() const;
-
-        /**
-         * Returns whether this route violates capacity constraints.
-         */
-        [[nodiscard]] bool hasExcessLoad() const;
-
-        /**
-         * Returns whether this route violates maximum distance constraints.
-         */
-        [[nodiscard]] bool hasExcessDistance() const;
-
-        /**
-         * Returns whether this route violates time window or maximum duration
-         * constraints.
-         */
-        [[nodiscard]] bool hasTimeWarp() const;
-
-        bool operator==(Route const &other) const;
-
-        Route() = delete;
-
-        Route(ProblemData const &data,
-              Visits visits,
-              VehicleType const vehicleType);
-
-        // This constructor does *no* validation. Useful when unserialising
-        // objects.
-        Route(Visits visits,
-              Distance distance,
-              Cost distanceCost,
-              Distance excessDistance,
-              Load delivery,
-              Load pickup,
-              Load excessLoad,
-              Duration duration,
-              Cost durationCost,
-              Duration timeWarp,
-              Duration travel,
-              Duration service,
-              Duration wait,
-              Duration release,
-              Duration startTime,
-              Duration slack,
-              Cost prizes,
-              std::pair<double, double> centroid,
-              VehicleType vehicleType,
-              Depot startDepot,
-              Depot endDepot);
-    };
-
-private:
     using Routes = std::vector<Route>;
     using Neighbours = std::vector<std::optional<std::pair<Client, Client>>>;
 
@@ -467,22 +237,12 @@ public:
      */
     Solution(ProblemData const &data, RandomNumberGenerator &rng);
 
-    /**
-     * Constructs a solution using routes given as lists of client indices.
-     * This constructor assumes all routes use vehicles having vehicle type 0.
-     *
-     * @param data   Data instance describing the problem that's being solved.
-     * @param routes Solution's route list.
-     */
+    // This constructs using routes given as lists of client indices. Assumes
+    // all routes use vehicles having vehicle type 0.
     Solution(ProblemData const &data,
              std::vector<std::vector<Client>> const &routes);
 
-    /**
-     * Constructs a solution from the given list of Routes.
-     *
-     * @param data   Data instance describing the problem that's being solved.
-     * @param routes Solution's route list.
-     */
+    // This constructs from the given list of Routes.
     Solution(ProblemData const &data, Routes const &routes);
 
     // This constructor does *no* validation. Useful when unserialising objects.
@@ -505,8 +265,6 @@ public:
 }  // namespace pyvrp
 
 std::ostream &operator<<(std::ostream &out, pyvrp::Solution const &sol);
-std::ostream &operator<<(std::ostream &out,
-                         pyvrp::Solution::Route const &route);
 
 template <> struct std::hash<pyvrp::Solution>
 {
