+++ pyvrp/cpp/repair/nearest_route_insert.cpp
@@ -10,14 +10,12 @@
 
 using pyvrp::Solution;
 using pyvrp::search::insertCost;
-using pyvrp::search::Route;
 
-using Locations = std::vector<Route::Node>;
-using Routes = std::vector<Route>;
-using SolRoutes = std::vector<Solution::Route>;
+using SearchRoute = pyvrp::search::Route;
+using SolRoute = pyvrp::Route;
 
-std::vector<Solution::Route>
-pyvrp::repair::nearestRouteInsert(SolRoutes const &solRoutes,
+std::vector<SolRoute>
+pyvrp::repair::nearestRouteInsert(std::vector<SolRoute> const &solRoutes,
                                   std::vector<size_t> const &unplanned,
                                   ProblemData const &data,
                                   CostEvaluator const &costEvaluator)
@@ -25,13 +23,13 @@ pyvrp::repair::nearestRouteInsert(SolRoutes const &solRoutes,
     if (solRoutes.empty() && !unplanned.empty())
         throw std::invalid_argument("Need routes to repair!");
 
-    Locations locs;
-    Routes routes;
+    std::vector<SearchRoute::Node> locs;
+    std::vector<SearchRoute> routes;
     setupRoutes(locs, routes, solRoutes, data);
 
     for (size_t client : unplanned)
     {
-        Route::Node *U = &locs[client];
+        SearchRoute::Node *U = &locs[client];
         ProblemData::Client const &clientData = data.location(client);
         assert(!U->route());
