+++ pyvrp/cpp/repair/nearest_route_insert.h
@@ -41,8 +41,8 @@ namespace pyvrp::repair
  *     When the list of routes is empty but the list of unplanned clients is
  *     not.
  */
-std::vector<Solution::Route>
-nearestRouteInsert(std::vector<Solution::Route> const &routes,
+std::vector<pyvrp::Route>
+nearestRouteInsert(std::vector<pyvrp::Route> const &routes,
                    std::vector<size_t> const &unplanned,
                    ProblemData const &data,
                    CostEvaluator const &costEvaluator);
